// client/src/graphql/user/__generated__/ChangePasswordAfterReset.ts
/* tslint:disable */
/* eslint-disable */
// @generated
// This file was automatically generated and should not be edited.

// ====================================================
// GraphQL mutation operation: ChangePasswordAfterReset
// ====================================================

export interface ChangePasswordAfterReset {
  changePasswordAfterReset: string | null;
}

export interface ChangePasswordAfterResetVariables {
  email: string;
  password: string;
  token: string;
}